# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/base/budget/create": {
            "post": {
                "description": "Create one Budget record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Create one Budget record",
                "parameters": [
                    {
                        "description": "Budget Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Budgets"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
...
        "/base/budget/one/{id}": {
            "get": {
                "description": "Get one budget record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Get one budget record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "budget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Budgets"
                        }
                    }
                }
            }
        },
        "/base/budget/update": {
            "post": {
                "description": "Update one budget record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Update one budget record",
                "parameters": [
                    {
                        "description": "Budget Request Body for update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Budgets"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
        "/base/budget/void/{id}": {
            "delete": {
                "description": "Delete one budget record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Delete one budget record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
...
        "/sys/budget/filter/list": {
            "post": {
                "description": "List Budgets With Filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "List Budgets With Filter",
                "parameters": [
                    {
                        "description": "List Budget Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ListBudgetRecordsDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Budgets"
                            }
                        }
                    }
                }
            }
        },
        "/sys/budget/list": {
            "post": {
                "description": "List Budgets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "List Budgets",
                "parameters": [
                    {
                        "description": "List Budget Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ListBudgetRecordsDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BudgetList"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BudgetList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Budgets"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GeneralCreateResponseBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateInactiveResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GeneralUpdateResponse"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateResponse": {
            "type": "object",
            "properties": {
                "MatchedCount": {
                    "type": "integer"
                },
                "ModifiedCount": {
                    "type": "integer"
                },
                "UpsertedCount": {
                    "type": "integer"
                },
                "UpsertedID": {
                    "type": "string"
                }
            }
        },
...
        "dto.ListBudgetRecordsDto": {
            "type": "object"
        },
...
        "models.Budgets": {
            "type": "object"
        },
...
    }
}